Add files via upload
this file contains 3 :  dashboard, analysis of the work done and the powerpoint showing the work with some analysis . 
By examining datasets on 
mental health, socio-economic indicators, and demographic 
trends, the project seeks to identify key predictors of suicide. 
The findings will be presented through an interactive 
dashboard, serving as a valuable tool for policymakers and 
mental health professionals.
The goal is to explore the dataset and provide some insights. 
Here are some suggestions: 
❏ Calculate the total number of suicides for each country 
and visualize the top 10 countries with the highest 
numbers.
 ❏ Analyze the trend of suicides over the years and 
visualize the global trend.
 ❏ Compare the suicide rates between males and females 
across different age groups.
 ❏ Investigate the relationship between GDP per capita 
and suicide rates using a scatter plot.
 ❏ Group the data by generation and compare the suicide 
rates to see if certain generations are more affected 
than others.